+++ game_patch/rf/physics.h
@@ -79,7 +79,7 @@ namespace rf
 
     static auto& gravity = addr_as_ref<float>(0x005A00DC);
     static auto& level_set_gravity = addr_as_ref<void(float value)>(0x004A0E20);
-
+    static auto& mp_ground_acceleration = addr_as_ref<float>(0x007C7084);
     static auto& physics_create_object = addr_as_ref<void(PhysicsData *pd, ObjectCreateInfo *oci)>(0x0049EC90);
     static auto& physics_delete_object = addr_as_ref<void(PhysicsData *pd)>(0x0049F1D0);
 }